Backend: add presets for individual sports
