+++ newswires/app/conf/Categories.scala
@@ -1,0 +1,1580 @@
+package conf
+
+private[conf] object CategoryCodes {
+  object Sport {
+    val AFP = "afpCat:SPO"
+    val AP = "apCat:s"
+    val PA = "paCat:SSS"
+  }
+
+  object Soccer {
+    val REUTERS = "N2:SOCC"
+    val PA = "paCat:SSO"
+    val AAP = "subj:15054000"
+  }
+
+  object Cricket {
+    val REUTERS = "N2:CRIC"
+    val PA = "paCat:SCR"
+    val AAP = "subj:15017000"
+  }
+
+  object RugbyUnion {
+    val REUTERS = "N2:RUGU"
+    val AAP = "subj:15049000"
+  }
+
+  object Tennis {
+    val REUTERS = "N2:TENN"
+    val AAP = "subj:15065000"
+  }
+
+  object Cycling {
+    val REUTERS = "N2:CYCL"
+    val AAP = "subj:15019000"
+  }
+
+  object F1 {
+    val REUTERS = "N2:FO1"
+    val AAP = "subj:15046001"
+  }
+
+  object Golf {
+    val REUTERS = "N2:GOLF"
+    val AAP = "subj:15027000"
+  }
+
+  object Boxing {
+    val REUTERS = "N2:BOXI"
+    val AAP = "subj:15014000"
+  }
+
+  object RugbyLeague {
+    val REUTERS = "N2:RUGL"
+    val AAP = "subj:15048000"
+  }
+
+  object Racing {
+    val REUTERS = "N2:HORS"
+    val PA = "paCat:SRR"
+    val AAP = "subj:15030000"
+  }
+}
+
+object Categories {
+  private[conf] val businessRelatedTopicCodes = List(
+    // Assets
+    "N2:ABS",
+    "N2:AGN",
+    "N2:BQ1",
+    "N2:CDM",
+    "N2:CDO",
+    "N2:CDV",
+    "N2:COM",
+    "N2:DBT",
+    "N2:DRV",
+    "N2:EQB",
+    "N2:ETF",
+    "N2:EUB",
+    "N2:FRX",
+    "N2:GOS",
+    "N2:GVD",
+    "N2:HOT",
+    "N2:HYD",
+    "N2:IGD",
+    "N2:ILB",
+    "N2:ILS",
+    "N2:LOA",
+    "N2:MMT",
+    "N2:MTG",
+    "N2:OPT",
+    "N2:PPL",
+    "N2:PVE",
+    "N2:REF",
+    "N2:SHL",
+    "N2:SOV",
+    "N2:STX",
+    "N2:SUK",
+    "N2:TA1",
+    "N2:TNC",
+    "N2:USC",
+    "N2:AUCT",
+    "N2:BINS",
+    "N2:CMBS",
+    "N2:COVB",
+    "N2:ETPS",
+    "N2:FUND",
+    "N2:MUNI",
+    "N2:NDFW",
+    "N2:PWRL",
+    "N2:REAM",
+    "N2:REVS",
+    "N2:SUPD",
+    "N2:CORPD",
+    "N2:HEDGE",
+    "N2:MTNOT",
+    "N2:RMBSE",
+    "N2:TWPWR",
+    "N2:BIOPWR",
+    "N2:COAPWR",
+    "N2:CRYCUR",
+    "N2:GASPWR",
+    "N2:GEOPWR",
+    "N2:HYDPWR",
+    "N2:NUCPWR",
+    "N2:OILPWR",
+    "N2:RNWPWR",
+    "N2:SOLPWR",
+    "N2:WINPWR",
+    // Commodities
+    "N2:ALU",
+    "N2:BUN",
+    "N2:CHR",
+    "N2:CHS",
+    "N2:CO2",
+    "N2:COA",
+    "N2:COC",
+    "N2:COF",
+    "N2:COR",
+    "N2:COT",
+    "N2:CRU",
+    "N2:GOL",
+    "N2:GRA",
+    "N2:H2O",
+    "N2:IRN",
+    "N2:JET",
+    "N2:LIV",
+    "N2:LNG",
+    "N2:MGS",
+    "N2:MLK",
+    "N2:MOG",
+    "N2:NAP",
+    "N2:NGL",
+    "N2:NGS",
+    "N2:NKL",
+    "N2:NRG",
+    "N2:ORJ",
+    "N2:PGM",
+    "N2:PWR",
+    "N2:RFO",
+    "N2:RNW",
+    "N2:RUB",
+    "N2:SLK",
+    "N2:STE",
+    "N2:SUG",
+    "N2:TEA",
+    "N2:TMP",
+    "N2:WEA",
+    "N2:WHT",
+    "N2:WND",
+    "N2:WOO",
+    "N2:ZNC",
+    "N2:AGRI",
+    "N2:ATMY",
+    "N2:BIOF",
+    "N2:BRGE",
+    "N2:BRLY",
+    "N2:BSMH",
+    "N2:CBLT",
+    "N2:CDMM",
+    "N2:CONT",
+    "N2:CPPR",
+    "N2:CTTL",
+    "N2:DAIR",
+    "N2:DIAM",
+    "N2:FERR",
+    "N2:FERT",
+    "N2:GALL",
+    "N2:GRMN",
+    "N2:HOIL",
+    "N2:INDI",
+    "N2:IRDM",
+    "N2:LITH",
+    "N2:LPG1",
+    "N2:MEAL",
+    "N2:METL",
+    "N2:MGSM",
+    "N2:MLDM",
+    "N2:NIOB",
+    "N2:OILS",
+    "N2:PALL",
+    "N2:PETC",
+    "N2:PHOS",
+    "N2:PLAS",
+    "N2:PLAT",
+    "N2:POIL",
+    "N2:POTH",
+    "N2:PRCP",
+    "N2:PROD",
+    "N2:RDFT",
+    "N2:RHDM",
+    "N2:RHEN",
+    "N2:RLFT",
+    "N2:RTNM",
+    "N2:RYE1",
+    "N2:SCRP",
+    "N2:SFTS",
+    "N2:SHFV",
+    "N2:SLCN",
+    "N2:SLNM",
+    "N2:SLVR",
+    "N2:SOIL",
+    "N2:SORG",
+    "N2:SOY1",
+    "N2:SSTE",
+    "N2:TGSN",
+    "N2:TIN1",
+    "N2:TMBR",
+    "N2:TNKR",
+    "N2:TNTE",
+    "N2:TTNM",
+    "N2:URAN",
+    "N2:VNDM",
+    "N2:AMCRU",
+    "N2:ASCRU",
+    "N2:BIOMS",
+    "N2:CTTLB",
+    "N2:DBULK",
+    "N2:LEAD1",
+    "N2:MECRU",
+    "N2:NSCRU",
+    "N2:PLTRY",
+    "N2:RAREE",
+    "N2:RICE1",
+    "N2:USCRU",
+    "N2:WINE1",
+    "N2:AFRCRU",
+    "N2:AUSCRU",
+    "N2:BASMTL",
+    "N2:BIOCEL",
+    "N2:BIODSL",
+    "N2:BIOETH",
+    "N2:CANCRU",
+    "N2:COCOIL",
+    "N2:COFARA",
+    "N2:COFROB",
+    "N2:COROIL",
+    "N2:DISTLL",
+    "N2:EMACRU",
+    "N2:LATCRU",
+    "N2:MINMTL",
+    "N2:NASCRU",
+    "N2:OLVOIL",
+    "N2:PIGHOG",
+    "N2:PNTOIL",
+    "N2:PORBEL",
+    "N2:PORPRO",
+    "N2:PREMTL",
+    "N2:RAPOIL",
+    "N2:RUSCRU",
+    "N2:SEACRU",
+    "N2:SNFOIL",
+    "N2:TRNSPT",
+    // General business news
+    "N2:AAA",
+    "N2:BL1",
+    "N2:CEN",
+    "N2:CM1",
+    "N2:CNH",
+    "N2:DIV",
+    "N2:ECI",
+    "N2:ESG",
+    "N2:GBR",
+    "N2:GDP",
+    "N2:INT",
+    "N2:IPO",
+    "N2:IPR",
+    "N2:ISU",
+    "N2:JOB",
+    "N2:MCE",
+    "N2:MRG",
+    "N2:NG1",
+    "N2:NT1",
+    "N2:PMI",
+    "N2:PRO",
+    "N2:PS1",
+    "N2:RCH",
+    "N2:RES",
+    "N2:RWS",
+    "N2:SL1",
+    "N2:STK",
+    "N2:ACCI",
+    "N2:BACT",
+    "N2:BKRT",
+    "N2:BONS",
+    "N2:BUYB",
+    "N2:CCRE",
+    "N2:CEO1",
+    "N2:CFO1",
+    "N2:CNSL",
+    "N2:DBTR",
+    "N2:DISP",
+    "N2:DVST",
+    "N2:FNAL",
+    "N2:ILEI",
+    "N2:INDX",
+    "N2:INFL",
+    "N2:MEVN",
+    "N2:MONS",
+    "N2:MPLT",
+    "N2:OSTR",
+    "N2:PINV",
+    "N2:PLCY",
+    "N2:PRIV",
+    "N2:PRXF",
+    "N2:QRND",
+    "N2:REGS",
+    "N2:RESF",
+    "N2:SHPP",
+    "N2:SISU",
+    "N2:STAT",
+    "N2:WPAY",
+    "N2:ALLCE",
+    "N2:BOSS1",
+    "N2:BSENT",
+    "N2:CAPUT",
+    "N2:CASE1",
+    "N2:CLASS",
+    "N2:CMPNY",
+    "N2:CPROD",
+    "N2:CSENT",
+    "N2:DDEAL",
+    "N2:DEAL1",
+    "N2:DEVAL",
+    "N2:DGOOD",
+    "N2:FINE1",
+    "N2:HOSAL",
+    "N2:INVTR",
+    "N2:KMOVE",
+    "N2:LIST1",
+    "N2:MEET1",
+    "N2:MONOP",
+    "N2:NAMEC",
+    "N2:PRES1",
+    "N2:QFNAL",
+    "N2:RECLL",
+    "N2:REORG",
+    "N2:REVAL",
+    "N2:SFNAL",
+    "N2:TRACC",
+    "N2:XPAND",
+    "N2:BKRFIG",
+    "N2:BNKCAP",
+    "N2:CARREG",
+    "N2:CHAIR1",
+    "N2:COMARB",
+    "N2:CONFIG",
+    "N2:CURINT",
+    "N2:LAYOFS",
+    "N2:MEDREG",
+    "N2:MNGISS",
+    "N2:MRTAPP",
+    "N2:ORDIMB",
+    "N2:RECAP1",
+    "N2:RESASS",
+    "N2:SHRACT",
+    "N2:SPLITB",
+    "N2:STSALE",
+    // Organisations
+    "N2:BBK",
+    "N2:BOC",
+    "N2:BOE",
+    "N2:BOJ",
+    "N2:ECB",
+    "N2:FED",
+    "N2:IMF",
+    "N2:RBA",
+    "N2:SEN",
+    "N2:SNB",
+    "N2:WTO",
+    "N2:AFRU",
+    "N2:APEC",
+    "N2:BMXO",
+    "N2:CBBR",
+    "N2:CBNG",
+    "N2:CBRT",
+    "N2:CBRU",
+    "N2:HNBK",
+    "N2:HREP",
+    "N2:NBPL",
+    "N2:OECD",
+    "N2:OPEC",
+    "N2:PBOC",
+    "N2:RBIN",
+    "N2:RBNZ",
+    "N2:SARB",
+    "N2:USDA",
+    "N2:WBNK",
+    "N2:NORBK",
+    "N2:RIKBK",
+    "N2:ASEAN1",
+    // Sectors
+    "N2:LUX",
+    "N2:MIN",
+    "N2:TMT",
+    "N2:AERO",
+    "N2:AGRC",
+    "N2:AIRL",
+    "N2:AIRP",
+    "N2:ALUM",
+    "N2:AMED",
+    "N2:APPA",
+    "N2:APRE",
+    "N2:AUTO",
+    "N2:BCST",
+    "N2:BEER",
+    "N2:BEVS",
+    "N2:BGDS",
+    "N2:BISV",
+    "N2:BMAT",
+    "N2:BNKS",
+    "N2:BSUP",
+    "N2:BSVC",
+    "N2:CARM",
+    "N2:CCOS",
+    "N2:CELE",
+    "N2:CHEM",
+    "N2:CHIP",
+    "N2:CIGS",
+    "N2:CMSS",
+    "N2:COAM",
+    "N2:COEN",
+    "N2:COFS",
+    "N2:COMA",
+    "N2:COMC",
+    "N2:COMS",
+    "N2:CONG",
+    "N2:CONS",
+    "N2:CYCP",
+    "N2:CYCS",
+    "N2:DEPT",
+    "N2:DISC",
+    "N2:DRIL",
+    "N2:ELCO",
+    "N2:ELEU",
+    "N2:ENER",
+    "N2:ENTS",
+    "N2:ENVS",
+    "N2:EXCA",
+    "N2:FARM",
+    "N2:FDRT",
+    "N2:FINS",
+    "N2:FOBE",
+    "N2:FOTB",
+    "N2:GASU",
+    "N2:HARW",
+    "N2:HECA",
+    "N2:HELE",
+    "N2:HFAC",
+    "N2:HMAC",
+    "N2:HMOS",
+    "N2:HOME",
+    "N2:HPAS",
+    "N2:HPRD",
+    "N2:HRSC",
+    "N2:INDG",
+    "N2:INDM",
+    "N2:INDS",
+    "N2:INSB",
+    "N2:INSR",
+    "N2:INVM",
+    "N2:INVT",
+    "N2:ISER",
+    "N2:ISLF",
+    "N2:ITSE",
+    "N2:LEIS",
+    "N2:LINS",
+    "N2:LIQS",
+    "N2:MACH",
+    "N2:MDIA",
+    "N2:MEDQ",
+    "N2:MEMI",
+    "N2:MINE",
+    "N2:MINS",
+    "N2:MRCH",
+    "N2:MTAL",
+    "N2:NCYC",
+    "N2:OFCE",
+    "N2:OGTR",
+    "N2:OILG",
+    "N2:OILI",
+    "N2:OILQ",
+    "N2:PACK",
+    "N2:PAPR",
+    "N2:PHAR",
+    "N2:PHMR",
+    "N2:PHON",
+    "N2:PINS",
+    "N2:PLAY",
+    "N2:PREC",
+    "N2:PRIN",
+    "N2:PSVC",
+    "N2:RDIY",
+    "N2:RDRL",
+    "N2:RDRU",
+    "N2:READ",
+    "N2:REAL",
+    "N2:REFI",
+    "N2:REIC",
+    "N2:REID",
+    "N2:REIR",
+    "N2:REIT",
+    "N2:RELX",
+    "N2:RENE",
+    "N2:RENF",
+    "N2:RENQ",
+    "N2:REOP",
+    "N2:RESV",
+    "N2:RETA",
+    "N2:RETE",
+    "N2:RETS",
+    "N2:RFOD",
+    "N2:RGEN",
+    "N2:RHOM",
+    "N2:RINS",
+    "N2:RRLF",
+    "N2:RRLP",
+    "N2:RUBB",
+    "N2:SBIZ",
+    "N2:SEMI",
+    "N2:SHOE",
+    "N2:SHOP",
+    "N2:SHOW",
+    "N2:SOFW",
+    "N2:STAY",
+    "N2:SWIT",
+    "N2:TCOM",
+    "N2:TECH",
+    "N2:TEEQ",
+    "N2:TRAD",
+    "N2:TRAN",
+    "N2:TREE",
+    "N2:TXTL",
+    "N2:UTIL",
+    "N2:UTIM",
+    "N2:WATR",
+    "N2:WEAR",
+    "N2:WLES",
+    "N2:WLTH",
+    "N2:WOOD",
+    "N2:ADVMK",
+    "N2:APRET",
+    "N2:AUTPT",
+    "N2:CASGM",
+    "N2:CHMSP",
+    "N2:CRDBX",
+    "N2:EXPRO",
+    "N2:FODPR",
+    "N2:HMFRN",
+    "N2:HRDWR",
+    "N2:HSEPR",
+    "N2:INVBR",
+    "N2:IRNST",
+    "N2:NPCNT",
+    "N2:PRTSV",
+    "N2:PSNPR",
+    "N2:REALE",
+    "N2:RECPR",
+    "N2:RESBR",
+    "N2:SEAFT",
+    "N2:SFTBV",
+    "N2:VHRET",
+    "N2:AIRFRT",
+    "N2:CHEMDV",
+    "N2:ENQSVC",
+    "N2:HLTHSE",
+    "N2:HSEBLD",
+    "N2:HSEGDS",
+    "N2:HSEHLD",
+    "N2:INVBIS",
+    "N2:INVDIV",
+    "N2:SHOPAL",
+    "N2:SPCRET",
+    "N2:NO EQUIVALENT"
+  )
+
+  private[conf] val sportRelatedTopicCodes = List(
+    "N2:DOP",
+    "N2:FO1",
+    "N2:MMA",
+    "N2:NBA",
+    "N2:NFL",
+    "N2:NHL",
+    "N2:SNO",
+    "N2:SPO",
+    "N2:ALPS",
+    "N2:Sports",
+    "N2:ANGL",
+    "N2:ARCH",
+    "N2:ATHL",
+    "N2:AUSR",
+    "N2:BADM",
+    "N2:BASE",
+    "N2:BASK",
+    "N2:BIAT",
+    "N2:BILL",
+    "N2:BOAR",
+    "N2:BOBS",
+    "N2:BOWL",
+    "N2:BOXI",
+    "N2:CANO",
+    "N2:CLMB",
+    "N2:CRIC",
+    "N2:CURL",
+    "N2:CYCL",
+    "N2:DIVE",
+    "N2:DNCG",
+    "N2:DRTS",
+    "N2:EQUE",
+    "N2:FENC",
+    "N2:FIGS",
+    "N2:FSKI",
+    "N2:GOLF",
+    "N2:GYMN",
+    "N2:HAND",
+    "N2:HOCK",
+    "N2:HORS",
+    "N2:ICEH",
+    "N2:JAIA",
+    "N2:JUDO",
+    "N2:KARA",
+    "N2:LACR",
+    "N2:LUGE",
+    "N2:MOCR",
+    "N2:MOCY",
+    "N2:MODE",
+    "N2:MORA",
+    "N2:MTHN",
+    "N2:NETB",
+    "N2:NORS",
+    "N2:RALL",
+    "N2:ROWI",
+    "N2:RUGL",
+    "N2:RUGU",
+    "N2:SDOG",
+    "N2:SHOO",
+    "N2:SKAT",
+    "N2:SKEL",
+    "N2:SKIJ",
+    "N2:SNOO",
+    "N2:SOCC",
+    "N2:SOFT",
+    "N2:SPEE",
+    "N2:SQUA",
+    "N2:STSK",
+    "N2:SUMO",
+    "N2:SWIM",
+    "N2:TABT",
+    "N2:TAEK",
+    "N2:TENN",
+    "N2:TENP",
+    "N2:TRIA",
+    "N2:VOLL",
+    "N2:WATP",
+    "N2:WATS",
+    "N2:WEIG",
+    "N2:WRES",
+    "N2:XCTY",
+    "N2:YACH",
+    "N2:AQSPO",
+    "N2:BVOLL",
+    "N2:CHESS",
+    "N2:MARTS",
+    "N2:ORENT",
+    "N2:PBOAT",
+    "N2:SBOAT",
+    "N2:SPOLO",
+    "N2:SPOOL",
+    "N2:SRFNG",
+    "N2:SYNCS"
+  )
+
+  private[conf] val sportsRelatedTopicCodes = List(
+    "N2:DOP",
+    "N2:FO1",
+    "N2:MMA",
+    "N2:NBA",
+    "N2:NFL",
+    "N2:NHL",
+    "N2:SNO",
+    "N2:SPO",
+    "N2:ALPS",
+    "N2:Sports",
+    "N2:ANGL",
+    "N2:ARCH",
+    "N2:ATHL",
+    "N2:AUSR",
+    "N2:BADM",
+    "N2:BASE",
+    "N2:BASK",
+    "N2:BIAT",
+    "N2:BILL",
+    "N2:BOAR",
+    "N2:BOBS",
+    "N2:BOWL",
+    "N2:BOXI",
+    "N2:CANO",
+    "N2:CLMB",
+    "N2:CRIC",
+    "N2:CURL",
+    "N2:CYCL",
+    "N2:DIVE",
+    "N2:DNCG",
+    "N2:DRTS",
+    "N2:EQUE",
+    "N2:FENC",
+    "N2:FIGS",
+    "N2:FSKI",
+    "N2:GOLF",
+    "N2:GYMN",
+    "N2:HAND",
+    "N2:HOCK",
+    "N2:HORS",
+    "N2:ICEH",
+    "N2:JAIA",
+    "N2:JUDO",
+    "N2:KARA",
+    "N2:LACR",
+    "N2:LUGE",
+    "N2:MOCR",
+    "N2:MOCY",
+    "N2:MODE",
+    "N2:MORA",
+    "N2:MTHN",
+    "N2:NETB",
+    "N2:NORS",
+    "N2:RALL",
+    "N2:ROWI",
+    "N2:RUGL",
+    "N2:RUGU",
+    "N2:SDOG",
+    "N2:SHOO",
+    "N2:SKAT",
+    "N2:SKEL",
+    "N2:SKIJ",
+    "N2:SNOO",
+    "N2:SOCC",
+    "N2:SOFT",
+    "N2:SPEE",
+    "N2:SQUA",
+    "N2:STSK",
+    "N2:SUMO",
+    "N2:SWIM",
+    "N2:TABT",
+    "N2:TAEK",
+    "N2:TENN",
+    "N2:TENP",
+    "N2:TRIA",
+    "N2:VOLL",
+    "N2:WATP",
+    "N2:WATS",
+    "N2:WEIG",
+    "N2:WRES",
+    "N2:XCTY",
+    "N2:YACH",
+    "N2:AQSPO",
+    "N2:BVOLL",
+    "N2:CHESS",
+    "N2:MARTS",
+    "N2:ORENT",
+    "N2:PBOAT",
+    "N2:SBOAT",
+    "N2:SPOLO",
+    "N2:SPOOL",
+    "N2:SRFNG",
+    "N2:SYNCS"
+  )
+
+  private[conf] val otherTopicCodes = List(
+    // Arts/Culture/Entertainment
+    "N2:ART",
+    "N2:BKS",
+    "N2:ENT",
+    "N2:FLM",
+    "N2:LIF",
+    "N2:CLEB",
+    "N2:FASH",
+    "N2:MSIC",
+    "N2:TELE",
+    "N2:THTR",
+    "N2:TOUR",
+    "N2:VGAME",
+    // Religion
+    "N2:BUD",
+    "N2:JUD",
+    "N2:REL",
+    "N2:CHRI",
+    "N2:HIND",
+    "N2:ISLM",
+    // Odd
+    "N2:ODD",
+    // Science/Technology
+    "N2:ENV",
+    "N2:GMO",
+    "N2:HEA",
+    "N2:SCI",
+    "N2:WWW",
+    "N2:AWLQ",
+    "N2:ITEC",
+    "N2:LSCI",
+    "N2:NATU",
+    "N2:SPEX"
+  )
+
+  private[conf] val businessRelatedNewsCode = List(
+    "subj:04004002",
+    "subj:04004003",
+    "subj:04004004",
+    "subj:04004005",
+    "subj:04004006",
+    "subj:04004007",
+    "subj:04005000",
+    "subj:04005001",
+    "subj:04005002",
+    "subj:04005003",
+    "subj:04005004",
+    "subj:04005005",
+    "subj:04005006",
+    "subj:04005007",
+    "subj:04005008",
+    "subj:04005009",
+    "subj:04005010",
+    "subj:04005011",
+    "subj:04005012",
+    "subj:04005013",
+    "subj:04005014",
+    "subj:04006000",
+    "subj:04006001",
+    "subj:04006002",
+    "subj:04006003",
+    "subj:04006004",
+    "subj:04006005",
+    "subj:04006006",
+    "subj:04006007",
+    "subj:04006008",
+    "subj:04006009",
+    "subj:04006010",
+    "subj:04006011",
+    "subj:04008000",
+    "subj:04008001",
+    "subj:04008002",
+    "subj:04008003",
+    "subj:04008004",
+    "subj:04008005",
+    "subj:04008006",
+    "subj:04008007",
+    "subj:04008008",
+    "subj:04008009",
+    "subj:04008010",
+    "subj:04008011",
+    "subj:04008012",
+    "subj:04008013",
+    "subj:04008014",
+    "subj:04008015",
+    "subj:04008016",
+    "subj:04008017",
+    "subj:04008018",
+    "subj:04008019",
+    "subj:04008020",
+    "subj:04008021",
+    "subj:04008022",
+    "subj:04008023",
+    "subj:04008024",
+    "subj:04008025",
+    "subj:04008026",
+    "subj:04008027",
+    "subj:04008028",
+    "subj:04008029",
+    "subj:04008030",
+    "subj:04008031",
+    "subj:04008032",
+    "subj:04008033",
+    "subj:04008034",
+    "subj:04008035",
+    "subj:04008036",
+    "subj:04009000",
+    "subj:04009001",
+    "subj:04009002",
+    "subj:04009003",
+    "subj:04009004",
+    "subj:04016000",
+    "subj:04016001",
+    "subj:04016002",
+    "subj:04016003",
+    "subj:04016004",
+    "subj:04016005",
+    "subj:04016006",
+    "subj:04016007",
+    "subj:04016008",
+    "subj:04016009",
+    "subj:04016010",
+    "subj:04016011",
+    "subj:04016012",
+    "subj:04016013",
+    "subj:04016014",
+    "subj:04016015",
+    "subj:04016016",
+    "subj:04016017",
+    "subj:04016018",
+    "subj:04016019",
+    "subj:04016020",
+    "subj:04016021",
+    "subj:04016022",
+    "subj:04016023",
+    "subj:04016024",
+    "subj:04016025",
+    "subj:04016026",
+    "subj:04016027",
+    "subj:04016028",
+    "subj:04016029",
+    "subj:04016030",
+    "subj:04016031",
+    "subj:04016032",
+    "subj:04016033",
+    "subj:04016034",
+    "subj:04016035",
+    "subj:04016036",
+    "subj:04016037",
+    "subj:04016038",
+    "subj:04016039",
+    "subj:04016040",
+    "subj:04016041",
+    "subj:04016042",
+    "subj:04016043",
+    "subj:04016044",
+    "subj:04016045",
+    "subj:04016046",
+    "subj:04016047",
+    "subj:04016048",
+    "subj:04016049",
+    "subj:04016050",
+    "subj:04016051",
+    "subj:04016052",
+    "subj:04016053",
+    "subj:04016054",
+    "subj:04016055",
+    "subj:04016056",
+    "subj:04016057",
+    "subj:04017000",
+    "subj:04017001",
+    "subj:04018000",
+    "subj:04018001",
+    "subj:04019000",
+    "subj:04019001",
+    "subj:04000000",
+    "subj:09000000",
+    "subj:09001000",
+    "subj:09002000",
+    "subj:09002001",
+    "subj:09002002",
+    "subj:09002003",
+    "subj:09003000",
+    "subj:09003001",
+    "subj:09003002",
+    "subj:09003003",
+    "subj:09003004",
+    "subj:09004000",
+    "subj:09005000",
+    "subj:09006000",
+    "subj:09007000",
+    "subj:09008000",
+    "subj:09009000",
+    "subj:09010000",
+    "subj:09011000",
+    "subj:09011001",
+    "subj:09011002",
+    "subj:09012000",
+    "subj:09013000",
+    "subj:09014000",
+    "subj:09015000",
+    "subj:09016000"
+  )
+
+  private[conf] val sportsRelatedNewsCodes = List(
+    "subj:15000000",
+    "subj:15001000",
+    "subj:15001001",
+    "subj:15001002",
+    "subj:15002000",
+    "subj:15002001",
+    "subj:15002002",
+    "subj:15002003",
+    "subj:15002004",
+    "subj:15002005",
+    "subj:15003000",
+    "subj:15003001",
+    "subj:15003002",
+    "subj:15003003",
+    "subj:15004000",
+    "subj:15004001",
+    "subj:15004002",
+    "subj:15005000",
+    "subj:15005001",
+    "subj:15005002",
+    "subj:15005003",
+    "subj:15005004",
+    "subj:15005005",
+    "subj:15005006",
+    "subj:15005007",
+    "subj:15005008",
+    "subj:15005009",
+    "subj:15005010",
+    "subj:15005011",
+    "subj:15005012",
+    "subj:15005013",
+    "subj:15005014",
+    "subj:15005015",
+    "subj:15005016",
+    "subj:15005017",
+    "subj:15005018",
+    "subj:15005019",
+    "subj:15005020",
+    "subj:15005021",
+    "subj:15005022",
+    "subj:15005023",
+    "subj:15005024",
+    "subj:15005025",
+    "subj:15005026",
+    "subj:15005027",
+    "subj:15005028",
+    "subj:15005029",
+    "subj:15005030",
+    "subj:15005031",
+    "subj:15005032",
+    "subj:15005033",
+    "subj:15005034",
+    "subj:15005035",
+    "subj:15005036",
+    "subj:15005037",
+    "subj:15005038",
+    "subj:15005039",
+    "subj:15005040",
+    "subj:15005041",
+    "subj:15005042",
+    "subj:15005043",
+    "subj:15005044",
+    "subj:15005045",
+    "subj:15005046",
+    "subj:15005047",
+    "subj:15005048",
+    "subj:15005049",
+    "subj:15005050",
+    "subj:15005051",
+    "subj:15005052",
+    "subj:15005053",
+    "subj:15005054",
+    "subj:15005055",
+    "subj:15005056",
+    "subj:15005057",
+    "subj:15005058",
+    "subj:15005059",
+    "subj:15005060",
+    "subj:15005061",
+    "subj:15005062",
+    "subj:15005063",
+    "subj:15005064",
+    "subj:15005065",
+    "subj:15005066",
+    "subj:15005067",
+    "subj:15005068",
+    "subj:15005069",
+    "subj:15006000",
+    "subj:15007000",
+    "subj:15007001",
+    "subj:15007002",
+    "subj:15007003",
+    "subj:15007004",
+    "subj:15007005",
+    "subj:15007006",
+    "subj:15008000",
+    "subj:15008001",
+    "subj:15008002",
+    "subj:15008003",
+    "subj:15008004",
+    "subj:15008005",
+    "subj:15009000",
+    "subj:15009001",
+    "subj:15009002",
+    "subj:15009003",
+    "subj:15009004",
+    "subj:15009005",
+    "subj:15009006",
+    "subj:15010000",
+    "subj:15010001",
+    "subj:15010002",
+    "subj:15010003",
+    "subj:15010004",
+    "subj:15010005",
+    "subj:15010006",
+    "subj:15011000",
+    "subj:15011001",
+    "subj:15011002",
+    "subj:15012000",
+    "subj:15013000",
+    "subj:15014000",
+    "subj:15014001",
+    "subj:15014002",
+    "subj:15014003",
+    "subj:15014004",
+    "subj:15014005",
+    "subj:15014006",
+    "subj:15014007",
+    "subj:15014008",
+    "subj:15014009",
+    "subj:15014010",
+    "subj:15014011",
+    "subj:15014012",
+    "subj:15014013",
+    "subj:15014014",
+    "subj:15014015",
+    "subj:15014016",
+    "subj:15014017",
+    "subj:15014018",
+    "subj:15014019",
+    "subj:15014020",
+    "subj:15014021",
+    "subj:15014022",
+    "subj:15014023",
+    "subj:15014024",
+    "subj:15015000",
+    "subj:15015001",
+    "subj:15015002",
+    "subj:15015003",
+    "subj:15015004",
+    "subj:15015005",
+    "subj:15015006",
+    "subj:15015007",
+    "subj:15015008",
+    "subj:15015009",
+    "subj:15015010",
+    "subj:15015011",
+    "subj:15015012",
+    "subj:15016000",
+    "subj:15016001",
+    "subj:15016002",
+    "subj:15017000",
+    "subj:15018000",
+    "subj:15018001",
+    "subj:15019000",
+    "subj:15019001",
+    "subj:15019002",
+    "subj:15019003",
+    "subj:15019004",
+    "subj:15019005",
+    "subj:15019006",
+    "subj:15019007",
+    "subj:15019008",
+    "subj:15019009",
+    "subj:15019010",
+    "subj:15019011",
+    "subj:15019012",
+    "subj:15019013",
+    "subj:15019014",
+    "subj:15019015",
+    "subj:15019016",
+    "subj:15019017",
+    "subj:15019018",
+    "subj:15019019",
+    "subj:15019020",
+    "subj:15019021",
+    "subj:15020000",
+    "subj:15021000",
+    "subj:15021001",
+    "subj:15021002",
+    "subj:15021003",
+    "subj:15021004",
+    "subj:15021005",
+    "subj:15021006",
+    "subj:15022000",
+    "subj:15022001",
+    "subj:15022002",
+    "subj:15022003",
+    "subj:15022004",
+    "subj:15023000",
+    "subj:15023001",
+    "subj:15023002",
+    "subj:15023003",
+    "subj:15024000",
+    "subj:15024001",
+    "subj:15025000",
+    "subj:15025001",
+    "subj:15025002",
+    "subj:15025003",
+    "subj:15026000",
+    "subj:15026001",
+    "subj:15026002",
+    "subj:15026003",
+    "subj:15027000",
+    "subj:15028000",
+    "subj:15028001",
+    "subj:15028002",
+    "subj:15028003",
+    "subj:15028004",
+    "subj:15028005",
+    "subj:15028006",
+    "subj:15028007",
+    "subj:15028008",
+    "subj:15028009",
+    "subj:15028010",
+    "subj:15028011",
+    "subj:15028012",
+    "subj:15028013",
+    "subj:15028014",
+    "subj:15028015",
+    "subj:15029000",
+    "subj:15030000",
+    "subj:15030001",
+    "subj:15030002",
+    "subj:15030003",
+    "subj:15030004",
+    "subj:15031000",
+    "subj:15031001",
+    "subj:15031002",
+    "subj:15032000",
+    "subj:15032001",
+    "subj:15032002",
+    "subj:15032003",
+    "subj:15032004",
+    "subj:15032005",
+    "subj:15032006",
+    "subj:15032007",
+    "subj:15032008",
+    "subj:15032009",
+    "subj:15032010",
+    "subj:15032011",
+    "subj:15032012",
+    "subj:15033000",
+    "subj:15033001",
+    "subj:15033002",
+    "subj:15033003",
+    "subj:15033004",
+    "subj:15033005",
+    "subj:15033006",
+    "subj:15033007",
+    "subj:15034000",
+    "subj:15034001",
+    "subj:15034002",
+    "subj:15035000",
+    "subj:15036000",
+    "subj:15036001",
+    "subj:15036002",
+    "subj:15037000",
+    "subj:15038000",
+    "subj:15038001",
+    "subj:15038002",
+    "subj:15038003",
+    "subj:15038004",
+    "subj:15038005",
+    "subj:15039000",
+    "subj:15039001",
+    "subj:15039002",
+    "subj:15039003",
+    "subj:15039004",
+    "subj:15039005",
+    "subj:15039006",
+    "subj:15039007",
+    "subj:15039008",
+    "subj:15040000",
+    "subj:15040001",
+    "subj:15040002",
+    "subj:15040003",
+    "subj:15041000",
+    "subj:15041001",
+    "subj:15041002",
+    "subj:15041003",
+    "subj:15041004",
+    "subj:15041005",
+    "subj:15041006",
+    "subj:15041007",
+    "subj:15041008",
+    "subj:15041009",
+    "subj:15041010",
+    "subj:15041011",
+    "subj:15041012",
+    "subj:15041013",
+    "subj:15041014",
+    "subj:15042000",
+    "subj:15043000",
+    "subj:15043001",
+    "subj:15043002",
+    "subj:15043003",
+    "subj:15043004",
+    "subj:15043005",
+    "subj:15043006",
+    "subj:15043007",
+    "subj:15043008",
+    "subj:15043009",
+    "subj:15043010",
+    "subj:15043011",
+    "subj:15043012",
+    "subj:15043013",
+    "subj:15043014",
+    "subj:15043015",
+    "subj:15043016",
+    "subj:15043017",
+    "subj:15044000",
+    "subj:15044001",
+    "subj:15045000",
+    "subj:15046000",
+    "subj:15046001",
+    "subj:15046002",
+    "subj:15047000",
+    "subj:15047001",
+    "subj:15047002",
+    "subj:15047003",
+    "subj:15047004",
+    "subj:15047005",
+    "subj:15047006",
+    "subj:15047007",
+    "subj:15048000",
+    "subj:15049000",
+    "subj:15049001",
+    "subj:15050000",
+    "subj:15050001",
+    "subj:15050002",
+    "subj:15050003",
+    "subj:15050004",
+    "subj:15050005",
+    "subj:15050006",
+    "subj:15050007",
+    "subj:15050008",
+    "subj:15050009",
+    "subj:15050010",
+    "subj:15050011",
+    "subj:15050012",
+    "subj:15050013",
+    "subj:15050014",
+    "subj:15050015",
+    "subj:15050016",
+    "subj:15051000",
+    "subj:15051001",
+    "subj:15051002",
+    "subj:15051003",
+    "subj:15051004",
+    "subj:15051005",
+    "subj:15051006",
+    "subj:15051007",
+    "subj:15051008",
+    "subj:15051009",
+    "subj:15051010",
+    "subj:15051011",
+    "subj:15051012",
+    "subj:15052000",
+    "subj:15052001",
+    "subj:15052002",
+    "subj:15052003",
+    "subj:15053000",
+    "subj:15053001",
+    "subj:15053002",
+    "subj:15054000",
+    "subj:15055000",
+    "subj:15056000",
+    "subj:15056001",
+    "subj:15056002",
+    "subj:15056003",
+    "subj:15056004",
+    "subj:15056005",
+    "subj:15056006",
+    "subj:15056007",
+    "subj:15056008",
+    "subj:15056009",
+    "subj:15056010",
+    "subj:15056011",
+    "subj:15056012",
+    "subj:15056013",
+    "subj:15056014",
+    "subj:15057000",
+    "subj:15058000",
+    "subj:15058001",
+    "subj:15058002",
+    "subj:15058003",
+    "subj:15058004",
+    "subj:15058005",
+    "subj:15059000",
+    "subj:15060000",
+    "subj:15061000",
+    "subj:15062000",
+    "subj:15062001",
+    "subj:15062002",
+    "subj:15062003",
+    "subj:15062004",
+    "subj:15062005",
+    "subj:15062006",
+    "subj:15062007",
+    "subj:15062008",
+    "subj:15062009",
+    "subj:15062010",
+    "subj:15062011",
+    "subj:15062012",
+    "subj:15062013",
+    "subj:15062014",
+    "subj:15062015",
+    "subj:15062016",
+    "subj:15062017",
+    "subj:15062018",
+    "subj:15062019",
+    "subj:15062020",
+    "subj:15062021",
+    "subj:15062022",
+    "subj:15062023",
+    "subj:15062024",
+    "subj:15062025",
+    "subj:15062026",
+    "subj:15063000",
+    "subj:15064000",
+    "subj:15064001",
+    "subj:15064002",
+    "subj:15064003",
+    "subj:15064004",
+    "subj:15064005",
+    "subj:15064006",
+    "subj:15064007",
+    "subj:15064008",
+    "subj:15065000",
+    "subj:15065001",
+    "subj:15066000",
+    "subj:15066001",
+    "subj:15066002",
+    "subj:15066003",
+    "subj:15067000",
+    "subj:15067001",
+    "subj:15068000",
+    "subj:15069000",
+    "subj:15069001",
+    "subj:15069002",
+    "subj:15069003",
+    "subj:15069004",
+    "subj:15070000",
+    "subj:15070001",
+    "subj:15070002",
+    "subj:15070003",
+    "subj:15070004",
+    "subj:15070005",
+    "subj:15070006",
+    "subj:15070007",
+    "subj:15070008",
+    "subj:15070009",
+    "subj:15070010",
+    "subj:15070011",
+    "subj:15070012",
+    "subj:15070013",
+    "subj:15070014",
+    "subj:15070015",
+    "subj:15070016",
+    "subj:15071000",
+    "subj:15071001",
+    "subj:15071002",
+    "subj:15071003",
+    "subj:15071004",
+    "subj:15072000",
+    "subj:15072001",
+    "subj:15072002",
+    "subj:15072003",
+    "subj:15072004",
+    "subj:15072005",
+    "subj:15072006",
+    "subj:15072007",
+    "subj:15072008",
+    "subj:15072009",
+    "subj:15072010",
+    "subj:15072011",
+    "subj:15072012",
+    "subj:15073000",
+    "subj:15073001",
+    "subj:15073002",
+    "subj:15073003",
+    "subj:15073004",
+    "subj:15073005",
+    "subj:15073006",
+    "subj:15073007",
+    "subj:15073008",
+    "subj:15073009",
+    "subj:15073010",
+    "subj:15073011",
+    "subj:15073012",
+    "subj:15073013",
+    "subj:15073014",
+    "subj:15073015",
+    "subj:15073016",
+    "subj:15073017",
+    "subj:15073018",
+    "subj:15073019",
+    "subj:15073020",
+    "subj:15073021",
+    "subj:15073022",
+    "subj:15073023",
+    "subj:15073024",
+    "subj:15073025",
+    "subj:15073026",
+    "subj:15073027",
+    "subj:15073028",
+    "subj:15073029",
+    "subj:15073030",
+    "subj:15073031",
+    "subj:15073032",
+    "subj:15073033",
+    "subj:15073034",
+    "subj:15073035",
+    "subj:15073036",
+    "subj:15073037",
+    "subj:15073038",
+    "subj:15073039",
+    "subj:15073040",
+    "subj:15073041",
+    "subj:15073042",
+    "subj:15073043",
+    "subj:15073044",
+    "subj:15073045",
+    "subj:15073046",
+    "subj:15073047",
+    "subj:15074000",
+    "subj:15074001",
+    "subj:15074002",
+    "subj:15074003",
+    "subj:15074004",
+    "subj:15074005",
+    "subj:15074006",
+    "subj:15074007",
+    "subj:15075000",
+    "subj:15076000",
+    "subj:15077000",
+    "subj:15077001",
+    "subj:15077002",
+    "subj:15077003",
+    "subj:15077004",
+    "subj:15077005",
+    "subj:15077006",
+    "subj:15077007",
+    "subj:15077008",
+    "subj:15077009",
+    "subj:15077010",
+    "subj:15078000",
+    "subj:15079000",
+    "subj:15080000",
+    "subj:15081000",
+    "subj:15082000",
+    "subj:15082001",
+    "subj:15082002",
+    "subj:15083000",
+    "subj:15084000",
+    "subj:15085000",
+    "subj:15086000",
+    "subj:15087000",
+    "subj:15088000",
+    "subj:15089000",
+    "subj:15090000",
+    "subj:15091000",
+    "subj:15092000",
+    "subj:15093000",
+    "subj:15094000",
+    "subj:15095000",
+    "subj:15096000",
+    "subj:15097000",
+    "subj:15098000",
+    "subj:15099000",
+    "subj:15100000",
+    "subj:15101000",
+    "subj:15102000",
+    "subj:15103000"
+  )
+}
